+++ src/lib/Containers/index.tsx
@@ -3,6 +3,7 @@ import Conversation from "./Conversation"
 import Gene from "./Gene"
 import Home from "./Home"
 import Inbox from "./Inbox"
+import Inquiry from "./Inquiry"
 import MyAccount from "./MyAccount"
 import WorksForYou from "./WorksForYou"
 
@@ -14,4 +15,5 @@ export default {
   WorksForYou,
   MyAccount,
   Inbox,
+  Inquiry,
 }